VACMS-12614: Refactor Cypress log-specific code out of `commands.js`
