+++ tests/cypress/support/commands.js
@@ -144,6 +144,7 @@ Cypress.Commands.add("type_ckeditor", (element, content) => {
     if (elements.indexOf(element) === -1) {
       const matches = elements.filter((el) => el.includes(element));
       if (matches.length) {
+        // eslint-disable-next-line prefer-destructuring
         element = matches[0];
       }
     }
@@ -243,25 +244,3 @@ Cypress.Commands.add("accessibilityLog", (violations) => {
 });
 
 compareSnapshotCommand();
-
-let logText = "";
-
-beforeEach(() => {
-  const testTitle = Cypress.currentTest.title;
-  const testPath = Cypress.currentTest.titlePath;
-  const date = new Date().toUTCString();
-  const timestamp = Math.floor(Date.now() / 1000);
-  logText += `VA_GOV_DEBUG ${timestamp} ${date} BEFORE ${testPath} ${testTitle}\n`;
-});
-
-afterEach(() => {
-  const testTitle = Cypress.currentTest.title;
-  const testPath = Cypress.currentTest.titlePath;
-  const date = new Date().toUTCString();
-  const timestamp = Math.floor(Date.now() / 1000);
-  logText += `VA_GOV_DEBUG ${timestamp} ${date} AFTER ${testPath} ${testTitle}\n`;
-});
-
-after(() => {
-  cy.writeFile("cypress.log", logText, { flag: "a+" });
-});
