+++ tests/cypress/support/index.js
@@ -14,3 +14,4 @@
 // ***********************************************************
 
 import "./commands";
+import "./log";
